fix: corrige bugs em avaliação e comentário-resposta, ajusta controllers e documenta serviços
Este commit resolve problemas relacionados à avaliação e à comentário-resposta, ajusta controllers e services e documenta as alterações feitas.

- Corrige bug na avaliação:
  - Resolve o erro no Swagger que impedia a inclusão do resenhaId.

- Corrige bug no comentário-resposta:
  - Resolve o erro no Swagger que impedia a inclusão do comentárioId.

- Ajustes nos controllers e services:
  - Realiza melhorias nos controllers e services de Avaliação e ComentárioResposta para corrigir os erros acima.

- Atualiza a documentação dos serviços:
  - Refatora a documentação para refletir as correções e melhorias realizadas nos serviços de Avaliação e ComentárioResposta.

+++ src/main/java/prati/projeto/redeSocial/service/impl/ComentarioRespostaServiceImpl.java
@@ -54,9 +54,13 @@ public class ComentarioRespostaServiceImpl implements ComentarioRespostaService 
 
     @Override
     @Transactional
-    public RespostaDTO atualizarResposta(Integer respostaId, RespostaDTO respostaDTO) {
+    public RespostaDTO atualizarResposta(Integer comentarioId, Integer respostaId, RespostaDTO respostaDTO) {
         ComentarioResposta resposta = respostaRepository.findById(respostaId)
                 .orElseThrow(() -> new RegraNegocioException("Resposta não encontrada"));
+
+        if (!resposta.getComentarioOriginal().getId().equals(comentarioId)) {
+            throw new RegraNegocioException("A resposta não pertence ao comentário informado.");
+        }
 
         resposta.setTexto(respostaDTO.getTexto());
         resposta.setDataResposta(LocalDateTime.now());
@@ -67,16 +71,25 @@ public class ComentarioRespostaServiceImpl implements ComentarioRespostaService 
 
     @Override
     @Transactional
-    public void deletarResposta(Integer respostaId) {
-        respostaRepository.findById(respostaId)
+    public void deletarResposta(Integer comentarioId, Integer respostaId) {
+        ComentarioResposta resposta = respostaRepository.findById(respostaId)
                 .orElseThrow(() -> new RegraNegocioException("Resposta não encontrada"));
+
+        if (!resposta.getComentarioOriginal().getId().equals(comentarioId)) {
+            throw new RegraNegocioException("A resposta não pertence ao comentário informado.");
+        }
+
         respostaRepository.deleteById(respostaId);
     }
 
     @Override
-    public RespostaDTO buscarRespostaPorId(Integer respostaId) {
+    public RespostaDTO buscarRespostaPorId(Integer comentarioId, Integer respostaId) {
         ComentarioResposta resposta = respostaRepository.findById(respostaId)
                 .orElseThrow(() -> new RegraNegocioException("Resposta não encontrada"));
+
+        if (!resposta.getComentarioOriginal().getId().equals(comentarioId)) {
+            throw new RegraNegocioException("A resposta não pertence ao comentário informado.");
+        }
 
         return convertToDTO(resposta);
     }
